Изменение типа первичных ключей на BIGINT
- Обновление скриптов миграций БД
- Обновление моделей

+++ reference/src/main/java/com/ebito/reference/repository/OperationRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface OperationRepository extends JpaRepository<Operation, Integer> {
-    List<Operation> findAllByClientId(int clientId);
+    List<Operation> findAllByClientId(long clientId);
 }
